Update patch set 1
Patch Set 1: Code-Review-1

Martin Welss is part of Acumos project and have the right to made change on acumos-c-client repo. Martin is the developper who will create on-boarding acumos client for models build with C or C++ language.

Regarding the license it's a mistake it will be corrected in the next edit of this patch

Patch-set: 1
Reviewer: Gerrit User 212 <212@655300a2-7a56-4504-9153-881edd907bea>
Label: Code-Review=-1